+++ src/LGML/Source/SerialController.cpp
@@ -50,7 +50,7 @@ void SerialController::buildLocalEnv() {
 	DynamicObject obj;
 	static const Identifier jsSendMessageIdentifier("sendMessage");
 	obj.setMethod(jsSendMessageIdentifier, sendMessageFromScript);
-	obj.setProperty(jsPtrIdentifier, this);
+	obj.setProperty(jsPtrIdentifier, (int64)this);
 
 	setLocalNamespace(obj);
 }
